requirements.txt
tensorflow==1.4.1
pandas==0.22.0
google-cloud-storage==1.6.0
